Move global hooks into hooks directory
