+++ site/source/components/PageData.tsx
@@ -4,15 +4,15 @@ import { useLocation } from 'react-router-dom'
 import styled from 'styled-components'
 
 import Meta from '@/components/utils/Meta'
-import { useIsEmbedded } from '@/components/utils/useIsEmbedded'
 import useSearchParamsSimulationSharing from '@/components/utils/useSearchParamsSimulationSharing'
-import useSimulationConfig from '@/components/utils/useSimulationConfig'
 import { Chip } from '@/design-system'
 import { Emoji } from '@/design-system/emoji'
 import { Spacing } from '@/design-system/layout'
 import { H1 } from '@/design-system/typography/heading'
 import { Intro } from '@/design-system/typography/paragraphs'
 import { useCurrentSimulatorData } from '@/hooks/useCurrentSimulatorData'
+import { useIsEmbedded } from '@/hooks/useIsEmbedded'
+import useSimulationConfig from '@/hooks/useSimulationConfig'
 import { situationSelector } from '@/store/selectors/simulationSelectors'
 
 import { NextSteps } from '../pages/Simulateurs/NextSteps'
