+++ site/source/components/conversation/AnswerList.tsx
@@ -6,7 +6,6 @@ import { useDispatch, useSelector } from 'react-redux'
 import styled from 'styled-components'
 
 import { EvaluatedRule, useEngine } from '@/components/utils/EngineContext'
-import { useNextQuestions } from '@/components/utils/useNextQuestion'
 import { Message, PopoverWithTrigger } from '@/design-system'
 import { Button } from '@/design-system/buttons'
 import { Emoji } from '@/design-system/emoji'
@@ -17,6 +16,7 @@ import { H2, H3 } from '@/design-system/typography/heading'
 import { Link } from '@/design-system/typography/link'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
 import { useCurrentSimulatorData } from '@/hooks/useCurrentSimulatorData'
+import { useNextQuestions } from '@/hooks/useNextQuestion'
 import { answerQuestion, resetSimulation } from '@/store/actions/actions'
 import { resetCompany } from '@/store/actions/companyActions'
 import { isCompanyDottedName } from '@/store/reducers/companySituationReducer'
