+++ site/source/pages/Simulateurs/Home.tsx
@@ -2,7 +2,6 @@ import { Trans, useTranslation } from 'react-i18next'
 
 import { TrackPage } from '@/components/ATInternetTracking'
 import PageHeader from '@/components/PageHeader'
-import { useIsEmbedded } from '@/components/utils/useIsEmbedded'
 import { Chip } from '@/design-system'
 import InfoBulle from '@/design-system/InfoBulle'
 import { Card } from '@/design-system/card'
@@ -15,6 +14,7 @@ import { Link } from '@/design-system/typography/link'
 import { Li, Ul } from '@/design-system/typography/list'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
 import { MergedSimulatorDataValues } from '@/hooks/useCurrentSimulatorData'
+import { useIsEmbedded } from '@/hooks/useIsEmbedded'
 import useSimulatorsData from '@/hooks/useSimulatorsData'
 import { useSitePaths } from '@/sitePaths'
