+++ site/source/pages/gerer/declaration-charges-sociales-independant/index.tsx
@@ -12,12 +12,12 @@ import RuleInput from '@/components/conversation/RuleInput'
 import Warning from '@/components/ui/WarningBlock'
 import { FromTop } from '@/components/ui/animate'
 import DefaultHelmet from '@/components/utils/DefaultHelmet'
-import useSimulationConfig from '@/components/utils/useSimulationConfig'
 import { Grid } from '@/design-system/layout'
 import { Strong } from '@/design-system/typography'
 import { H2, H3 } from '@/design-system/typography/heading'
 import { Li, Ul } from '@/design-system/typography/list'
 import { Body, Intro, SmallBody } from '@/design-system/typography/paragraphs'
+import useSimulationConfig from '@/hooks/useSimulationConfig'
 import { useSitePaths } from '@/sitePaths'
 import { updateSituation } from '@/store/actions/actions'
 import { SimulationConfig } from '@/store/reducers/rootReducer'
