+++ site/source/pages/gerer/declaration-revenu-independants/cotisations.tsx
@@ -5,12 +5,12 @@ import Progress from '@/components/ui/Progress'
 import { FromTop } from '@/components/ui/animate'
 import { useEngine } from '@/components/utils/EngineContext'
 import { Markdown } from '@/components/utils/markdown'
-import { useSimulationProgress } from '@/components/utils/useNextQuestion'
 import { Message } from '@/design-system'
 import { Container, Grid, Spacing } from '@/design-system/layout'
 import { Strong } from '@/design-system/typography'
 import { H2, H3 } from '@/design-system/typography/heading'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
+import { useSimulationProgress } from '@/hooks/useNextQuestion'
 
 import { SimpleField } from '../components/Fields'
 import { DéclarationRevenu } from './components/DéclarationRevenu'
