+++ site/source/pages/gerer/declaration-revenu-independants/index.tsx
@@ -5,10 +5,10 @@ import { Condition } from '@/components/EngineValue'
 import PrintExportRecover from '@/components/simulationExplanation/PrintExportRecover'
 import DefaultHelmet from '@/components/utils/DefaultHelmet'
 import { useEngine } from '@/components/utils/EngineContext'
-import { useSimulationProgress } from '@/components/utils/useNextQuestion'
-import useSimulationConfig from '@/components/utils/useSimulationConfig'
 import { Step, Stepper } from '@/design-system'
 import { Spacing } from '@/design-system/layout'
+import { useSimulationProgress } from '@/hooks/useNextQuestion'
+import useSimulationConfig from '@/hooks/useSimulationConfig'
 import { useSitePaths } from '@/sitePaths'
 import { SimulationConfig } from '@/store/reducers/rootReducer'
 import { omit } from '@/utils'
